Pull in the `utz` library for time zones
I cannot get `localtime` to work on the Pi Pico W. I've seen references
to it on the Pi forums, and I think other people might have had success,
but the libc we're using doesn't know what to do on a microcontroller.
That's fine, let's use a tiny one that only does what I need it to do
instead.